+++ data/tr3/ship/data/scripts/crash.lua
@@ -1,0 +1,3 @@
+trx.events.before_level_file(function(level)
+  trx.creatures.add_ally(trx.catalog.objects.sthpac_mercenary)
+end)